+++ infrastructure/standalone/iam.tf
@@ -1,0 +1,99 @@
+data "aws_iam_policy_document" "ec2_assume_role" {
+  statement {
+    actions = ["sts:AssumeRole"]
+    principals {
+      type        = "Service"
+      identifiers = ["ec2.amazonaws.com"]
+    }
+  }
+}
+
+resource "aws_iam_role" "application" {
+  name               = "${var.name_prefix}-instance"
+  assume_role_policy = data.aws_iam_policy_document.ec2_assume_role.json
+}
+
+resource "aws_iam_role_policy_attachment" "ssm" {
+  role       = aws_iam_role.application.name
+  policy_arn = "arn:aws:iam::aws:policy/AmazonSSMManagedInstanceCore"
+}
+
+data "aws_iam_policy_document" "resume_storage" {
+  statement {
+    actions   = ["s3:DeleteObject", "s3:GetObject", "s3:PutObject"]
+    resources = ["${aws_s3_bucket.resumes.arn}/resumes/*"]
+  }
+
+  statement {
+    actions   = ["ssm:GetParameter"]
+    resources = ["arn:aws:ssm:${var.aws_region}:${data.aws_caller_identity.current.account_id}:parameter/${var.name_prefix}/deploy/*"]
+  }
+}
+
+resource "aws_iam_role_policy" "resume_storage" {
+  name   = "PrivateResumeStorage"
+  role   = aws_iam_role.application.id
+  policy = data.aws_iam_policy_document.resume_storage.json
+}
+
+resource "aws_iam_instance_profile" "application" {
+  name = "${var.name_prefix}-instance"
+  role = aws_iam_role.application.name
+}
+
+resource "aws_iam_openid_connect_provider" "github" {
+  url             = "https://token.actions.githubusercontent.com"
+  client_id_list  = ["sts.amazonaws.com"]
+  thumbprint_list = ["ab9d0263244dd0326eb67015705a667e79cfe998"]
+}
+
+data "aws_iam_policy_document" "github_assume_role" {
+  statement {
+    actions = ["sts:AssumeRoleWithWebIdentity"]
+    principals {
+      type        = "Federated"
+      identifiers = [aws_iam_openid_connect_provider.github.arn]
+    }
+    condition {
+      test     = "StringEquals"
+      variable = "token.actions.githubusercontent.com:aud"
+      values   = ["sts.amazonaws.com"]
+    }
+    condition {
+      test     = "StringEquals"
+      variable = "token.actions.githubusercontent.com:sub"
+      values   = [var.github_oidc_subject]
+    }
+  }
+}
+
+resource "aws_iam_role" "github_deploy" {
+  name               = "${var.name_prefix}-github-deploy"
+  assume_role_policy = data.aws_iam_policy_document.github_assume_role.json
+}
+
+data "aws_iam_policy_document" "github_deploy" {
+  statement {
+    actions = ["ssm:SendCommand"]
+    resources = [
+      aws_instance.application.arn,
+      "arn:aws:ssm:${var.aws_region}::document/AWS-RunShellScript",
+    ]
+  }
+  statement {
+    actions   = ["ssm:GetCommandInvocation"]
+    resources = ["*"]
+  }
+  statement {
+    actions = ["ssm:PutParameter", "ssm:DeleteParameter"]
+    resources = [
+      "arn:aws:ssm:${var.aws_region}:${data.aws_caller_identity.current.account_id}:parameter/${var.name_prefix}/deploy/*",
+    ]
+  }
+}
+
+resource "aws_iam_role_policy" "github_deploy" {
+  name   = "DeployThroughSsm"
+  role   = aws_iam_role.github_deploy.id
+  policy = data.aws_iam_policy_document.github_deploy.json
+}